pyproject.toml
[project]
name = "image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aspose-words>=25.11.0",
    "pillow>=12.0.0",
]
